requirements.txt
Django==3.2.12
django_autoslug==1.9.8
django_email_verification==0.3.1
django_environ==0.10.0
python-decouple==3.8
python_dateutil==2.8.1
